Set default local_dns_server to 114.114.114.114
If using 127.0.0.1, meanwhile the proxy server address is a domain that needs to be resolved to IP address, connections routing to proxy will cause V2ray resolving by itself which means the DNS query will be redirected to the "domain'ed" proxy over and over again til it timed out.

Thus, we recommended using 114.114.114.114 as local_dns_server, just in case.