Support enabling a DHCP server powered by dnsmasq for special routers.
Because like the default firmware for Xiaomi routers, there's couple of routers that not support changing DHCP options.  We couldn't send NanoPi as the gateway to clients. So that maintain a DHCP server by ourself and then turn off the built-in one may be the only solution without having to flash another firmware.